requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
pyyaml==6.0.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
celery[redis]==5.5.2
grpcio==1.73.1
grpcio-tools==1.73.1
httpx==0.28.1
pyjwt==2.10.1
